Move things around and restore logout
