+++ app/src/main/java/alison/fivethingskotlin/ContainerActivity.kt
@@ -4,7 +4,11 @@ import alison.fivethingskotlin.Fragments.AnalyticsFragment
 import alison.fivethingskotlin.Fragments.DesignsFragment
 import alison.fivethingskotlin.Fragments.FiveThingsFragment
 import alison.fivethingskotlin.Fragments.SettingsFragment
+import alison.fivethingskotlin.Util.Constants.ACCOUNT_TYPE
+import alison.fivethingskotlin.Util.Constants.AUTH_TOKEN_TYPE
+import android.accounts.AccountManager
 import android.content.Context
+import android.content.Intent
 import android.os.Bundle
 import android.support.design.widget.NavigationView
 import android.support.v4.view.GravityCompat
@@ -15,6 +19,7 @@ import android.support.v7.widget.Toolbar
 import android.view.MenuItem
 import android.view.View
 import uk.co.chrisjenx.calligraphy.CalligraphyContextWrapper
+
 
 class ContainerActivity : AppCompatActivity() {
 
@@ -92,6 +97,10 @@ class ContainerActivity : AppCompatActivity() {
                     loadFragment(SettingsFragment())
                     true
                 }
+                R.id.logout_item -> {
+                    logOut()
+                    true
+                }
                 else -> {
                     true
                 }
@@ -107,6 +116,18 @@ class ContainerActivity : AppCompatActivity() {
         fragmentTransaction.replace(R.id.content_frame, fragment)
         fragmentTransaction.commitAllowingStateLoss()
         drawerLayout.closeDrawers()
+    }
+
+    private fun logOut() {
+        val accountManager = AccountManager.get(this)
+        val accounts = accountManager.getAccountsByType(ACCOUNT_TYPE)
+        val authToken = accountManager.peekAuthToken(accounts[0], AUTH_TOKEN_TYPE)
+        accountManager.invalidateAuthToken(ACCOUNT_TYPE, authToken)
+        accountManager.removeAccount(accounts[0], PromoActivity(), null, null)
+
+        val intent = Intent(applicationContext, PromoActivity::class.java)
+        intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
+        startActivity(intent)
     }
 
 }
