+++ app/src/main/res/layout/activity_promo.xml
@@ -53,7 +53,7 @@
             android:layout_marginStart="16dp"
             android:background="@color/bluegreen"
             android:padding="6dp"
-            android:text="Sign Up"
+            android:text="Sign In"
             android:textColor="@color/white"
             app:layout_constraintBottom_toTopOf="@+id/createAccountButton"
             app:layout_constraintEnd_toEndOf="parent"
