+++ app/src/main/res/menu/drawer_view.xml
@@ -17,5 +17,10 @@
             android:id="@+id/settings_item"
             android:icon="@drawable/ic_settings"
             android:title="@string/settings" />
+
+        <item
+            android:id="@+id/logout_item"
+            android:icon="@drawable/ic_signout"
+            android:title="@string/log_out" />
     </group>
 </menu>
