+++ app/src/main/res/values/strings.xml
@@ -9,4 +9,5 @@
     <string name="five_things">Five Things</string>
     <string name="settings">Settings</string>
     <string name="already_have_an_account">Already have an account?</string>
+    <string name="log_out">Log out</string>
 </resources>
